+++ CHANGELOG.md
@@ -15,7 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added ARRAYFORMULA function. (#630)
 
 ### Fixed
-- Fixed an issue with arrays and cruds. (#651)
+- Fixed an issue with useArrayArithmetic and cruds. (#651)
 
 ## [0.6.0] - 2021-04-27